+++ tests/Simple.Logging.Console.Tests/AnsiParserTests.cs
@@ -1,4 +1,4 @@
-namespace Crude.Logging.Console.Tests;
+namespace Simple.Logging.Console.Tests;
 
 public class AnsiParserTests
 {